+++ ErsatzTV.Application/Channels/Commands/UpdateChannelHandler.cs
@@ -43,7 +43,6 @@ public class UpdateChannelHandler(
         c.MusicVideoCreditsMode = update.MusicVideoCreditsMode;
         c.MusicVideoCreditsTemplate = update.MusicVideoCreditsTemplate;
         c.SongVideoMode = update.SongVideoMode;
-        c.ActiveMode = update.ActiveMode;
         c.TranscodeMode = update.TranscodeMode;
         c.IdleBehavior = update.IdleBehavior;
         c.IsEnabled = update.IsEnabled;