+++ ErsatzTV.Infrastructure.MySql/Migrations/20250804211606_Remove_ChannelActiveMode.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ErsatzTV.Infrastructure.MySql.Migrations
+{
+    /// <inheritdoc />
+    public partial class Remove_ChannelActiveMode : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "ActiveMode",
+                table: "Channel");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "ActiveMode",
+                table: "Channel",
+                type: "int",
+                nullable: false,
+                defaultValue: 0);
+        }
+    }
+}