+++ CHANGELOG.md
@@ -74,6 +74,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - This controls the progression of the channel's playout, and has nothing to do with transcoding
   - `Always` is now called `Continuous` (playout progresses with wall clock)
   - `On Demand` is unchanged (playout only progresses while a client is watching the channel)
+- Replace channel `Active Mode` setting with new `Is Enabled` and `Show In EPG` settings
+  - `Active` channels will be converted to `Is Enabled` = true and `Show In EPG` = true
+  - `Hidden` channels will be converted to `Is Enabled` = true and `Show In EPG` = false
+  - `Inactive` channels will be converted to `Is Enabled` = false and `Show In EPG` = false
 
 ## [25.3.1] - 2025-07-24
 ### Fixed